requirements.txt
altair==4.2.2
attrs==22.2.0
blinker==1.5
cachetools==5.3.0
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
exceptiongroup==1.1.1
filelock==3.10.4
fonttools==4.39.2
geographiclib==2.0
geopy==2.3.0
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
imageio==2.26.1
importlib-metadata==6.1.0
importlib-resources==5.12.0
iniconfig==2.0.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.0
nltk==3.8.1
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
plotly==5.13.1
pluggy==1.0.0
protobuf==3.20.3
pyarrow==11.0.0
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.2.2
python-dateutil==2.8.2
pytz==2023.2
pytz-deprecation-shim==0.1.0.post0
regex==2023.3.23
requests==2.28.2
rich==13.3.2
river==0.15.0
scipy==1.10.1
seaborn==0.12.2
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.20.0
streamlit-plotly-events==0.0.6
sympy==1.11.1
tenacity==8.2.2
tensorboard-plugin-wit==1.8.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.0
tornado==6.2
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.2
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
watchdog==3.0.0
wincertstore==0.2
zipp==3.15.0
